Aggiunto redirect alla pagina di login
git-svn-id: svn+ssh://svnselezioni.si.cnr.it/DATA/svnroot/svn-documentale/branches/development/cool-no-surf@6315 18059cec-d5c9-4679-bb3d-60c028e053f1

+++ cool-common/cool-common-rest/src/main/java/it/cnr/cool/rest/Content.java
@@ -6,14 +6,17 @@ import it.cnr.cool.security.SecurityChecked;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.net.URI;
+import java.net.URISyntaxException;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.Context;
+import javax.ws.rs.core.Response;
+import javax.ws.rs.core.Response.ResponseBuilder;
 
 import org.apache.chemistry.opencmis.client.api.Document;
 import org.apache.chemistry.opencmis.client.api.Session;
@@ -36,27 +39,26 @@ public class Content {
 	private static final Logger LOGGER = LoggerFactory.getLogger(Content.class);
 
 	@GET
-	public void content(@Context HttpServletRequest req,
+	public Response content(@Context HttpServletRequest req,
 			@Context HttpServletResponse res, @QueryParam("path") String path,
 			@QueryParam("nodeRef") String nodeRef,
-			@QueryParam("deleteAfterDownload") Boolean deleteAfterDownload) {
-
+			@QueryParam("deleteAfterDownload") Boolean deleteAfterDownload) throws URISyntaxException {
 		Session cmisSession = cmisService.getCurrentCMISSession(req);
 
-		Document document;
-
-		if (path != null && !path.isEmpty()) {
-			String cleanPath = path.replaceAll("\\?.*", "");
-			LOGGER.debug("get content for path: " + cleanPath);
-			document = (Document) cmisSession.getObjectByPath(cleanPath);
-		} else {
-			LOGGER.debug("get content for nodeRef: " + nodeRef);
-			document = (Document) cmisSession.getObject(nodeRef);
-		}
-
-		res.setContentType(document.getContentStreamMimeType());
-
+		Document document = null;
 		try {
+
+			if (path != null && !path.isEmpty()) {
+				String cleanPath = path.replaceAll("\\?.*", "");
+				LOGGER.debug("get content for path: " + cleanPath);
+				document = (Document) cmisSession.getObjectByPath(cleanPath);
+			} else {
+				LOGGER.debug("get content for nodeRef: " + nodeRef);
+				document = (Document) cmisSession.getObject(nodeRef);
+			}
+
+			res.setContentType(document.getContentStreamMimeType());
+
 			String attachFileName = document.getContentStreamFileName();
 			String headerValue = "attachment";
 			if (attachFileName != null && !attachFileName.isEmpty()) {
@@ -77,18 +79,30 @@ public class Content {
 			inputStream.close();
 			outputStream.close();
 		} catch(CmisUnauthorizedException e) {
-            LOGGER.info("unauthorized to get " + nodeRef);
-			res.setStatus(HttpStatus.SC_UNAUTHORIZED);
+            LOGGER.debug("unauthorized to get " + nodeRef);
+            String redirect = "/" + Page.LOGIN_URL;
+            redirect = redirect.concat("?redirect=rest/content");
+			if (path != null && !path.isEmpty())
+				redirect = redirect.concat("&path="+path);
+			if (nodeRef != null && !nodeRef.isEmpty())
+				redirect = redirect.concat("&nodeRef="+nodeRef);
+			return Response.seeOther(new URI(getUrl(req) + redirect)).build();
 		} catch (IOException e) {
 			LOGGER.error("unable to get content for path " + path, e);
 			res.setStatus(HttpStatus.SC_INTERNAL_SERVER_ERROR);
 		}
 
-		if (deleteAfterDownload != null && deleteAfterDownload) {
+		if (document != null && deleteAfterDownload != null && deleteAfterDownload) {
 			LOGGER.info("deleting " + document.getName() + " (" + document.getId() + ")");
 			document.delete();
 		}
+		return Response.ok().build();
+	}
 
+	static String getUrl(HttpServletRequest req) {
+		StringBuffer url = req.getRequestURL();
+		int l = url.indexOf(req.getServletPath());
+		return url.substring(0, l);
 	}
 
 }
